Incorporate feedback: Mostly split up the Http class
  * Split up Http class
  * Set UserAgent when creating the HttpClient bean instead of scattered
    throughout the code.
