+++ rpki-validator/src/main/java/net/ripe/rpki/validator3/rrdp/HttpRrdpClient.java
@@ -32,14 +32,13 @@ package net.ripe.rpki.validator3.rrdp;
 import lombok.extern.slf4j.Slf4j;
 import net.ripe.rpki.validator3.api.util.BuildInformation;
 import net.ripe.rpki.validator3.domain.metrics.HttpClientMetricsService;
-import net.ripe.rpki.validator3.util.Http;
+import net.ripe.rpki.validator3.util.http.HttpStreaming;
 import org.eclipse.jetty.client.HttpClient;
 import org.eclipse.jetty.client.api.Request;
 import org.eclipse.jetty.http.HttpHeader;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import javax.annotation.PostConstruct;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -68,11 +67,9 @@ public class HttpRrdpClient implements RrdpClient {
         long before = System.currentTimeMillis();
         String statusDescription = "200";
         try {
-            return Http.readStream(() -> {
+            return HttpStreaming.readStream(() -> {
                 final Request request = httpClient.newRequest(uri);
                 final String version = buildInformation.getVersion();
-                request.header(HttpHeader.USER_AGENT, null);
-                request.header(HttpHeader.USER_AGENT, "RIPE NCC RPKI Validator/" + version);
                 return request;
             }, reader);
         } catch (Exception e) {
